Share the SSZ input builder between both generators
`build_stateless_input` and `empty_execution_requests` existed as two near-identical
copies, one per binary in this crate, differing only in how the block type was spelled.
The file that carried the second copy said a real PR should lift them into a module, and
this is that PR. They now live in src/lib.rs and both binaries import them, so the next
ethrex rev bump moves one encoder instead of two that nothing forced to stay in step.

Behaviour-preserving, and checked that way rather than assumed: after the move all five
generated fixtures still come out at the digests they already had, the real-block one
byte-for-byte at 08a52e10 with the same installed state root.
